General improvements
Optimised imports
Debloated boot.py
Added menus.py
Created themes folder
Removed demo
Added requirements.txt
Updated README
A few other small changes
