+++ pycos/README.md
@@ -1,1 +1,17 @@
 # PycOS source
+
+## Directory structure
+
+**`__init__.py` files have been excluded.**
+
+* system/
+    * gooey/
+        * manager.py
+        * menus.py
+    * recovery.py
+    * sysdata.py
+* themes/
+    * raspberry.py
+* boot.py
+* config.py
+* main.py
