--- pycos/src/pycos-installer.cmd
@@ -1,10 +1,0 @@
-rshell mkdir /pyboard//system
-rshell cp ./boot.py /pyboard/boot.py
-rshell cp ./config.py /pyboard/config.py
-rshell cp ./main.py /pyboard/main.py
-rshell mkdir /pyboard//system/gooey
-rshell cp ./system/recovery.py /pyboard/system/recovery.py
-rshell cp ./system/sysdata.py /pyboard/system/sysdata.py
-rshell cp ./system/__init__.py /pyboard/system/__init__.py
-rshell cp ./system/gooey/manager.py /pyboard/system/gooey/manager.py
-rshell cp ./system/gooey/__init__.py /pyboard/system/gooey/__init__.py
