+++ pycos/src/system/sysdata.py
@@ -1,2 +1,3 @@
-pycos_version = "0.0.2"
+pycos_version = "" # prerelease
+
 reset_causes = ["PWRON_RESET", "HARD_RESET", "WDT_RESET", "DEEPSLEEP_RESET", "SOFT_RESET"]
