+++ pycos/src/themes/raspberry.py
@@ -1,0 +1,17 @@
+BACKGROUND_COLOUR = (255, 0, 0) # colour in rgb
+# background colour
+# allowed values: tuple ranging from (0, 0, 0) to (255, 255, 255)
+# recommended values: any
+# default: (255, 0, 0)
+
+TEXT_COLOUR = (255, 255, 255) # colour in rgb
+# text colour
+# allowed values: tuple ranging from (0, 0, 0) to (255, 255, 255)
+# recommended values: any
+# default value: (255, 255, 255)
+
+SEPARATOR_COLOUR = (255, 85, 85) # colour in rgb
+# separator colour
+# allowed values: tuple ranging from (0, 0, 0) to (255, 255, 255)
+# recommended values: any
+# default value: (255, 85, 85)
