+++ requirements.txt
@@ -1,0 +1,2 @@
+rshell~=0.0.30
+mpy-cross~=1.16
